python_files/requirements.txt
pandas
mysql.connector
matplotlib
selenium
requests
webdriver-manager
elsapy